Allow specifying pipe syntax for use-ranges checks
Add `UseReversePipe` option to (boost|modernize)-use-ranges checks.
This controls whether to create a reverse view using function syntax
(`reverse(Range)`) or pipe syntax (`Range | reverse`)
